Undo bootstrap container
- Make new WithSearch trait
