+++ resources/views/bootstrap-4/datatable.blade.php
@@ -10,6 +10,7 @@
             wire:poll="{{ $refresh }}"
         @endif
     @endif
+    class="container-fluid"
 >
     @include('livewire-tables::bootstrap-4.includes.offline')
     @include('livewire-tables::bootstrap-4.includes.sorting-pills')
