+++ src/Traits/WithFilters.php
@@ -31,21 +31,6 @@ trait WithFilters
     ];
 
     /**
-     * @var int|null
-     */
-    public ?int $searchFilterDebounce = null;
-
-    /**
-     * @var bool|null
-     */
-    public ?bool $searchFilterDefer = null;
-
-    /**
-     * @var bool|null
-     */
-    public ?bool $searchFilterLazy = null;
-
-    /**
      * Prebuild the $filters array
      */
     public function mountWithFilters(): void
@@ -55,28 +40,6 @@ trait WithFilters
                 $this->filters[$filter] = null;
             }
         }
-    }
-
-    /**
-     * Build Livewire model options for the search input
-     *
-     * @return string
-     */
-    public function getSearchFilterOptionsProperty(): string
-    {
-        if ($this->searchFilterDebounce) {
-            return '.debounce.' . $this->searchFilterDebounce . 'ms';
-        }
-
-        if ($this->searchFilterDefer) {
-            return '.defer';
-        }
-
-        if ($this->searchFilterLazy) {
-            return '.lazy';
-        }
-
-        return '';
     }
 
     /**
